Analytic Yeni/calisma/requirements.txt
azure-iot-device~=2.10
pandas~=1.3.3
pyModbusTCP~=0.1.10


